+++ fluent-react/examples/async-messages/src/l10n.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import delay from 'delay';
 
 import 'fluent-intl-polyfill/compat';
-import { MessageContext } from 'fluent/compat';
+import { FluentBundle } from 'fluent/compat';
 import { LocalizationProvider } from 'fluent-react/compat';
 import { negotiateLanguages } from 'fluent-langneg/compat';
 
@@ -19,15 +19,15 @@ async function createMessagesGenerator(currentLocales) {
   const fetched = await Promise.all(
     currentLocales.map(fetchMessages)
   );
-  const bundle = fetched.reduce(
+  const messages = fetched.reduce(
     (obj, cur) => Object.assign(obj, cur)
   );
 
   return function* generateMessages() {
     for (const locale of currentLocales) {
-      const cx = new MessageContext(locale);
-      cx.addMessages(bundle[locale]);
-      yield cx;
+      const bundle = new FluentBundle(locale);
+      bundle.addMessages(messages[locale]);
+      yield bundle;
     }
   }
 }