+++ fluent-react/examples/fallback-sync/src/l10n.js
@@ -1,5 +1,5 @@
 import 'fluent-intl-polyfill/compat';
-import { MessageContext } from 'fluent/compat';
+import { FluentBundle } from 'fluent/compat';
 
 const MESSAGES_ALL = {
   'pl': `
@@ -13,8 +13,8 @@ bar = Bar in English
 
 export function* generateMessages() {
   for (const locale of ['pl', 'en-US']) {
-    const cx = new MessageContext(locale);
-    cx.addMessages(MESSAGES_ALL[locale]);
-    yield cx;
+    const bundle = new FluentBundle(locale);
+    bundle.addMessages(MESSAGES_ALL[locale]);
+    yield bundle;
   }
 }