+++ fluent-react/examples/higher-order/src/l10n.js
@@ -1,5 +1,5 @@
 import 'fluent-intl-polyfill/compat';
-import { MessageContext } from 'fluent/compat';
+import { FluentBundle } from 'fluent/compat';
 import { negotiateLanguages } from 'fluent-langneg/compat';
 
 const MESSAGES_ALL = {
@@ -22,8 +22,8 @@ export function* generateMessages(userLocales) {
   );
 
   for (const locale of currentLocales) {
-    const cx = new MessageContext(locale);
-    cx.addMessages(MESSAGES_ALL[locale]);
-    yield cx;
+    const bundle = new FluentBundle(locale);
+    bundle.addMessages(MESSAGES_ALL[locale]);
+    yield bundle;
   }
 }