+++ fluent-react/examples/redux-async/src/actions.js
@@ -1,4 +1,4 @@
-import { MessageContext } from 'fluent/compat';
+import { FluentBundle } from 'fluent/compat';
 import { negotiateLanguages } from 'fluent-langneg/compat';
 import delay from 'delay';
 
@@ -25,15 +25,15 @@ export function changeLocales(userLocales) {
       currentLocales.map(fetchMessages)
     );
 
-    const bundle = fetched.reduce(
+    const messages = fetched.reduce(
       (obj, cur) => Object.assign(obj, cur)
     );
 
     const generateMessages = function* () {
       for (const locale of currentLocales) {
-        const cx = new MessageContext(locale);
-        cx.addMessages(bundle[locale]);
-        yield cx;
+        const bundle = new FluentBundle(locale);
+        bundle.addMessages(messages[locale]);
+        yield bundle;
       }
     }