+++ fluent-sequence/src/map_sync.js
@@ -1,28 +1,28 @@
 /*
  * Synchronously map an identifier or an array of identifiers to the best
- * `MessageContext` instance(s).
+ * `FluentBundle` instance(s).
  *
  * @param {Iterable} iterable
  * @param {string|Array<string>} ids
- * @returns {MessageContext|Array<MessageContext>}
+ * @returns {FluentBundle|Array<FluentBundle>}
  */
 export default function mapContextSync(iterable, ids) {
   if (!Array.isArray(ids)) {
-    return getContextForId(iterable, ids);
+    return getBundleForId(iterable, ids);
   }
 
   return ids.map(
-    id => getContextForId(iterable, id)
+    id => getBundleForId(iterable, id)
   );
 }
 
 /*
- * Find the best `MessageContext` with the translation for `id`.
+ * Find the best `FluentBundle` with the translation for `id`.
  */
-function getContextForId(iterable, id) {
-  for (const context of iterable) {
-    if (context.hasMessage(id)) {
-      return context;
+function getBundleForId(iterable, id) {
+  for (const bundle of iterable) {
+    if (bundle.hasMessage(id)) {
+      return bundle;
     }
   }