+++ fluent/test/functions_builtin_test.js
@@ -2,16 +2,16 @@
 
 import assert from 'assert';
 
-import { MessageContext } from '../src/context';
+import { FluentBundle } from '../src/context';
 import { ftl } from '../src/util';
 
 suite('Built-in functions', function() {
-  let ctx;
+  let bundle;
 
   suite('NUMBER', function(){
     suiteSetup(function() {
-      ctx = new MessageContext('en-US', { useIsolating: false });
-      ctx.addMessages(ftl`
+      bundle = new FluentBundle('en-US', { useIsolating: false });
+      bundle.addMessages(ftl`
         num-decimal = { NUMBER($arg) }
         num-percent = { NUMBER($arg, style: "percent") }
         num-bad-opt = { NUMBER($arg, style: "bad") }
@@ -21,49 +21,49 @@ suite('Built-in functions', function() {
     test('missing argument', function() {
       let msg;
 
-      msg = ctx.getMessage('num-decimal');
-      assert.equal(ctx.format(msg), 'NaN');
+      msg = bundle.getMessage('num-decimal');
+      assert.equal(bundle.format(msg), 'NaN');
 
-      msg = ctx.getMessage('num-percent');
-      assert.equal(ctx.format(msg), 'NaN');
+      msg = bundle.getMessage('num-percent');
+      assert.equal(bundle.format(msg), 'NaN');
 
-      msg = ctx.getMessage('num-bad-opt');
-      assert.equal(ctx.format(msg), 'NaN');
+      msg = bundle.getMessage('num-bad-opt');
+      assert.equal(bundle.format(msg), 'NaN');
     });
 
     test('number argument', function() {
       const args = {arg: 1};
       let msg;
 
-      msg = ctx.getMessage('num-decimal');
-      assert.equal(ctx.format(msg, args), '1');
+      msg = bundle.getMessage('num-decimal');
+      assert.equal(bundle.format(msg, args), '1');
 
-      msg = ctx.getMessage('num-percent');
-      assert.equal(ctx.format(msg, args), '100%');
+      msg = bundle.getMessage('num-percent');
+      assert.equal(bundle.format(msg, args), '100%');
 
-      msg = ctx.getMessage('num-bad-opt');
-      assert.equal(ctx.format(msg, args), '1');
+      msg = bundle.getMessage('num-bad-opt');
+      assert.equal(bundle.format(msg, args), '1');
     });
 
     test('string argument', function() {
       const args = {arg: "Foo"};
       let msg;
 
-      msg = ctx.getMessage('num-decimal');
-      assert.equal(ctx.format(msg, args), 'NaN');
+      msg = bundle.getMessage('num-decimal');
+      assert.equal(bundle.format(msg, args), 'NaN');
 
-      msg = ctx.getMessage('num-percent');
-      assert.equal(ctx.format(msg, args), 'NaN');
+      msg = bundle.getMessage('num-percent');
+      assert.equal(bundle.format(msg, args), 'NaN');
 
-      msg = ctx.getMessage('num-bad-opt');
-      assert.equal(ctx.format(msg, args), 'NaN');
+      msg = bundle.getMessage('num-bad-opt');
+      assert.equal(bundle.format(msg, args), 'NaN');
     });
   });
 
   suite('DATETIME', function(){
     suiteSetup(function() {
-      ctx = new MessageContext('en-US', { useIsolating: false });
-      ctx.addMessages(ftl`
+      bundle = new FluentBundle('en-US', { useIsolating: false });
+      bundle.addMessages(ftl`
         dt-default = { DATETIME($arg) }
         dt-month = { DATETIME($arg, month: "long") }
         dt-bad-opt = { DATETIME($arg, month: "bad") }
@@ -73,14 +73,14 @@ suite('Built-in functions', function() {
     test('missing argument', function() {
       let msg;
 
-      msg = ctx.getMessage('dt-default');
-      assert.equal(ctx.format(msg), 'Invalid Date');
+      msg = bundle.getMessage('dt-default');
+      assert.equal(bundle.format(msg), 'Invalid Date');
 
-      msg = ctx.getMessage('dt-month');
-      assert.equal(ctx.format(msg), 'Invalid Date');
+      msg = bundle.getMessage('dt-month');
+      assert.equal(bundle.format(msg), 'Invalid Date');
 
-      msg = ctx.getMessage('dt-bad-opt');
-      assert.equal(ctx.format(msg), 'Invalid Date');
+      msg = bundle.getMessage('dt-bad-opt');
+      assert.equal(bundle.format(msg), 'Invalid Date');
     });
 
     test('Date argument', function () {
@@ -94,18 +94,18 @@ suite('Built-in functions', function() {
       const args = {arg: date};
       let msg;
 
-      msg = ctx.getMessage('dt-default');
-      assert.equal(ctx.format(msg, args), expectedDefault);
+      msg = bundle.getMessage('dt-default');
+      assert.equal(bundle.format(msg, args), expectedDefault);
 
-      msg = ctx.getMessage('dt-month');
-      assert.equal(ctx.format(msg, args), expectedMonth);
+      msg = bundle.getMessage('dt-month');
+      assert.equal(bundle.format(msg, args), expectedMonth);
 
-      msg = ctx.getMessage('dt-bad-opt');
+      msg = bundle.getMessage('dt-bad-opt');
       // The argument value will be coerced into a string by the join operation
-      // in MessageContext.format.  The result looks something like this; it
+      // in FluentBundle.format.  The result looks something like this; it
       // may vary depending on the TZ:
       //     Thu Sep 29 2016 02:00:00 GMT+0200 (CEST)
-      assert.equal(ctx.format(msg, args), date.toString());
+      assert.equal(bundle.format(msg, args), date.toString());
     });
   });
 });